refactor: Move the `schema` command to `debug schema`
