+++ influxdb_iox/src/commands/debug/mod.rs
@@ -1,12 +1,20 @@
-use snafu::{ResultExt, Snafu};
+use futures::Future;
+use influxdb_iox_client::connection::Connection;
+use snafu::prelude::*;
 
 mod dump_catalog;
 mod print_cpu;
+mod schema;
 
 #[derive(Debug, Snafu)]
 pub enum Error {
+    #[snafu(context(false))]
     #[snafu(display("Error in dump-catalog subcommand: {}", source))]
     DumpCatalogError { source: dump_catalog::Error },
+
+    #[snafu(context(false))]
+    #[snafu(display("Error in schema subcommand: {}", source))]
+    SchemaError { source: schema::Error },
 }
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
@@ -25,16 +33,23 @@ enum Command {
 
     /// Prints what CPU features are used by the compiler by default.
     PrintCpu,
+
+    Schema(schema::Config),
 }
 
-pub async fn command(config: Config) -> Result<()> {
+pub async fn command<C, CFut>(connection: C, config: Config) -> Result<()>
+where
+    C: Send + FnOnce() -> CFut,
+    CFut: Send + Future<Output = Connection>,
+{
     match config.command {
-        Command::DumpCatalog(dump_catalog) => dump_catalog::command(*dump_catalog)
-            .await
-            .context(DumpCatalogSnafu),
-        Command::PrintCpu => {
-            print_cpu::main();
-            Ok(())
+        Command::DumpCatalog(dump_catalog) => dump_catalog::command(*dump_catalog).await?,
+        Command::PrintCpu => print_cpu::main(),
+        Command::Schema(config) => {
+            let connection = connection().await;
+            schema::command(connection, config).await?
         }
     }
+
+    Ok(())
 }
